Move existing schemas to Hardship folder
