+++ maat-orchestration/src/main/java/uk/gov/justice/laa/crime/orchestration/mapper/HardshipMapper.java
@@ -5,9 +5,8 @@ import org.springframework.stereotype.Component;
 import uk.gov.justice.laa.crime.orchestration.dto.WorkflowRequest;
 import uk.gov.justice.laa.crime.orchestration.dto.maat.*;
 import uk.gov.justice.laa.crime.orchestration.enums.*;
-import uk.gov.justice.laa.crime.orchestration.model.ApiFindHardshipResponse;
-import uk.gov.justice.laa.crime.orchestration.model.ApiHardshipDetail;
-import uk.gov.justice.laa.crime.orchestration.model.ApiHardshipProgress;
+import uk.gov.justice.laa.crime.orchestration.model.court_data_api.hardship.ApiHardshipDetail;
+import uk.gov.justice.laa.crime.orchestration.model.court_data_api.hardship.ApiHardshipProgress;
 import uk.gov.justice.laa.crime.orchestration.model.hardship.*;
 import uk.gov.justice.laa.crime.orchestration.util.CurrencyUtil;
 import uk.gov.justice.laa.crime.orchestration.util.DateUtil;
