+++ maat-orchestration/src/main/resources/schemas/court-data-api/hardship/apiHardshipDetail.json
