+++ maat-orchestration/src/main/resources/schemas/hardship/apiFindHardshipResponse.json
@@ -47,7 +47,7 @@
     "solicitorCosts": {
       "type": "object",
       "description": "Solicitor cost",
-      "$ref": "solicitorCosts.json"
+      "$ref": "common/solicitorCosts.json"
     },
     "status": {
       "type": "object",
@@ -58,14 +58,14 @@
       "type": "array",
       "description": "Hardship review detail items",
       "items": {
-        "$ref": "apiHardshipDetail.json"
+        "$ref": "../court-data-api/hardship/apiHardshipDetail.json"
       }
     },
     "reviewProgressItems": {
       "type": "array",
       "description": "The hardship review progress details (actions taken & responses)",
       "items": {
-        "$ref": "apiHardshipProgress.json"
+        "$ref": "../court-data-api/hardship/apiHardshipProgress.json"
       }
     }
   },
