--- maat-orchestration/src/main/resources/schemas/solicitorCosts.json
@@ -1,30 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-04/schema#",
-  "id": "solicitorCosts.json",
-  "type": "object",
-  "title": "The Solicitor Costs schema",
-  "description": "Details of solicitor cost",
-  "properties": {
-    "rate": {
-      "type": "number",
-      "description": "The solicitors hourly rate"
-    },
-    "hours": {
-      "type": "integer",
-      "description": "The number of billable hours"
-    },
-    "vat": {
-      "type": "number",
-      "description": "Solicitors VAT costs"
-    },
-    "disbursements": {
-      "type": "number",
-      "description": "Solicitors disbursement costs"
-    },
-    "estimatedTotal": {
-      "type": "number",
-      "description": "The total estimated cost (this is automatically calculated in MAAT)"
-    }
-  },
-  "required": ["rate", "hours", "vat", "disbursements", "estimatedTotal"]
-}
